+++ CHANGELOG.md
@@ -2,15 +2,23 @@
 
 All notable changes to the "azureblockchain" extension will be documented in this file.
 
-## 0.1.0 
 
-- Initial release
+
+## 0.1.3
+
+- various bug fixes
+- moved logic app/function/flow generators out of Truffle build directory into their own directory
+- refactoring of Welcome/Requirements pages
+
+## 0.1.2
+
+- doc and bug fixes for contract code generation
 
 ## 0.1.1
 
 - Updated menu options
 
-## 0.1.2
+## 0.1.0 
 
-- doc and bug fixes for contract code generation
+- Initial release